Minor review and doc update

+++ EL_FitsFile/src/lib/MefFile.cpp
@@ -26,7 +26,7 @@ namespace FitsIO {
 
 MefFile::MefFile(std::string filename, Permission permission) :
     FitsFile(filename, permission),
-    m_hdus(std::max(long(1), Cfitsio::Hdu::count(m_fptr))) {} // 1 for create, count() for open
+    m_hdus(std::max(1L, Cfitsio::Hdu::count(m_fptr))) {} // 1 for create, count() for open
 
 long MefFile::hdu_count() const {
   return m_hdus.size();
