+++ EL_FitsIO_Validation/src/program/EL_FitsIO_BufferedReadPerf.cpp
@@ -36,9 +36,9 @@ using value_type = float;
 using column_type = VecColumn<value_type>;
 using table_type = std::vector<column_type>;
 
-table_type generate_table(int cols, int rows) {
+table_type generate_table(long cols, long rows) {
   table_type table(cols);
-  for (int c=0; c<cols; ++c) {
+  for (long c=0; c<cols; ++c) {
     auto data = Test::generate_random_vector<value_type>(rows);
     table[c] = column_type { { std::to_string(c), "", 1 }, std::move(data) };
   }
